chore: Remove duplicate logic from `~/lib/style.ts`
- We were defining the color roles in 2 places.
