+++ mobile/src/components/Form/Slider.tsx
@@ -19,8 +19,8 @@ import type { Icon } from "~/resources/icons/type";
 
 import { Colors } from "~/constants/Styles";
 import { OnRTL } from "~/lib/react";
-import type { AppColor } from "~/lib/style";
 import { cn } from "~/lib/style";
+import type { AppColor } from "~/modules/theme/constants";
 import { useColor } from "~/modules/theme/hooks";
 import { Em } from "../Typography/StyledText";
