+++ mobile/src/components/List/ListItemContent.tsx
@@ -3,8 +3,8 @@ import { memo } from "react";
 import { useTranslation } from "react-i18next";
 import { View } from "react-native";
 
-import type { VariantColor } from "~/lib/style";
 import { cn } from "~/lib/style";
+import type { VariantColor } from "~/modules/theme/constants";
 import { useTheme } from "~/modules/theme/hooks";
 import { StyledText } from "../Typography/StyledText";
