+++ mobile/src/components/Marquee.tsx
@@ -14,8 +14,8 @@ import { View } from "react-native";
 import { useInForeground } from "~/stores/ListenerState";
 
 import { OnRTLWorklet } from "~/lib/react";
-import type { ColorRole } from "~/lib/style";
 import { cn } from "~/lib/style";
+import type { ColorRole } from "~/modules/theme/constants";
 import { useColor } from "~/modules/theme/hooks";
 import { ScrollView } from "./Base/ScrollView";
