+++ mobile/src/lib/style.ts
@@ -4,40 +4,12 @@ import { extendTailwindMerge } from "tailwind-merge";
 
 import { BorderRadius, FontFamily, FontSize } from "~/constants/Styles";
 import { toLowerCase } from "~/utils/string";
+import { ColorRoleOptions } from "~/modules/theme/constants";
 import type { Font } from "~/stores/Preference/constants";
 
 //#region Color
-// We have some non-standard color roles:
-//  - `errorDim`, `onErrorVariant`
-//  - If we treat these as "fixed" colors, they're "standard" names:
-//    - `primaryDim`, `onPrimaryVariant`, `secondaryDim`, `onSecondaryVariant`
-const ColorRoles = [
-  ...["primary", "primaryDim", "onPrimary", "onPrimaryVariant"],
-  ...["secondary", "secondaryDim", "onSecondary", "onSecondaryVariant"],
-  ...["error", "errorDim", "onError", "onErrorVariant"],
-  ...["surfaceDim", "surface", "surfaceBright"],
-  ...[
-    "surfaceContainerLowest",
-    "surfaceContainerLow",
-    "surfaceContainer",
-    "surfaceContainerHigh",
-    "surfaceContainerHighest",
-  ],
-  ...["onSurface", "onSurfaceVariant", "outline", "outlineVariant"],
-  ...["inverseSurface", "inverseOnSurface"],
-] as const;
-
-export type ColorRole = (typeof ColorRoles)[number];
-export type HexColor = `#${string}`;
-export type AppColor = ColorRole | HexColor;
-/** List of colors which also has a `Variant` color. */
-export type VariantColor =
-  Extract<ColorRole, `${string}Variant`> extends `${infer Prefix}Variant`
-    ? Prefix
-    : never;
-
 // Need to include `transparent` as otherwise, things will get merged incorrectly.
-const AvailableColors = ["transparent", ...ColorRoles] as const;
+const AvailableColors = ["transparent", ...ColorRoleOptions] as const;
 
 const customTwMerge = extendTailwindMerge({
   override: {
@@ -55,11 +27,6 @@ const customTwMerge = extendTailwindMerge({
 /** Combines any number of Tailwind classes nicely. */
 export function cn(...inputs: ClassValue[]) {
   return customTwMerge(clsx(inputs));
-}
-
-/** Determines if a string is a hex color. */
-export function isHexColor(color?: string): color is HexColor {
-  return color !== undefined && color.startsWith("#");
 }
 //#endregion
