+++ mobile/src/modules/theme/screens/View.tsx
@@ -14,14 +14,13 @@ import { PreferenceSetters } from "~/stores/Preference/actions";
 import { PagePlaceholder } from "~/navigation/components/Placeholder";
 import { ScreenOptions } from "~/navigation/components/ScreenOptions";
 
-import type { HexColor } from "~/lib/style";
 import { cn } from "~/lib/style";
 import { FlatList } from "~/components/Base/List";
 import { Pressable } from "~/components/Base/Pressable";
 import { FilledIconButton, IconButton } from "~/components/Form/Button/Icon";
 import { StyledText } from "~/components/Typography/StyledText";
 import { exportTheme } from "../helpers/backup";
-import type { CustomTheme } from "../constants";
+import type { CustomTheme, HexColor } from "../constants";
 import {
   DefaultThemeOptions,
   Themes as DefaultThemes,
