+++ mobile/src/resources/icons/type.ts
@@ -1,4 +1,4 @@
-import type { AppColor } from "~/lib/style";
+import type { AppColor } from "~/modules/theme/constants";
 
 export type Icon = {
   /** Defaults to `24px`. */
